% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calib_unit_test}
\alias{calib_unit_test}
\title{Results from calibration study for one serological test}
\usage{
calib_unit_test()
}
\description{
Results from calibration study for one serological test
}
