% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{code_region}
\alias{code_region}
\title{code_region}
\format{
A data frame of 324 categories of 4 variables:
\describe{
  \item{REGION}{code region INSEE (18)}
  \item{CHEFLIEU}{Code de la commune chef-lieu}
  \item{TNCC}{Type de nom (article)}
  \item{NCC}{Nom en clair (majuscules)}
  \item{NCCENR}{Nom en clair (typographie riche)}
  \item{reg_spf, REG_SPF1, REG_SPF2, reg_spf2}{Variations on region labels with lower/upper case and accents}
  \item{abb}{abbreviated labels}
  }
}
\source{
\url{https://www.data.gouv.fr/fr/datasets/r/6585c878-4e2f-4ceb-a219-75f6631d1f14}
}
\usage{
code_region
}
\description{
Code and labels of French region 
(Code officiel géographique au 1er janvier 2016)
}
\keyword{datasets}
