% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mci_2var}
\alias{mci_2var}
\title{Summarize iterations stratified by two categorical variables
For each period}
\usage{
mci_2var(v1, v2, pop = pops, df = psw, time = "week")
}
\arguments{
\item{v1}{categorical variable}

\item{v2}{categorical variable 2}

\item{pop}{population counts}

\item{df}{data.frame of poststratification cells + period}

\item{time}{period}
}
\description{
Summarize iterations stratified by two categorical variables
For each period
}
