% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mci_var}
\alias{mci_var}
\title{Summarize iterations stratified by one categorical variable
For each period}
\usage{
mci_var(v = "age", pop = pops, df = psw, time = "week")
}
\arguments{
\item{v}{categorical variable}

\item{pop}{population counts}

\item{df}{data.frame of poststratification cells + period}

\item{time}{period}
}
\description{
Summarize iterations stratified by one categorical variable
For each period
}
