% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{poststrat}
\alias{poststrat}
\title{poststrat}
\format{
A data frame of 324 categories of 4 variables:
\describe{
  \item{code}{INSEE code of region (18)}
  \item{sex}{Sex: "Homme" "Femme"}
  \item{age}{Age class: "0-9"   "10-19" "20-29" "30-39" "40-49" "50-59" "60-69" "70-79" "80+"}
  \item{n}{Population size}
  }
}
\source{
\url{https://www.insee.fr/fr/statistiques/1893198}
}
\usage{
poststrat
}
\description{
Population counts by strata of sex, age, region.
Use insee pop counts as of 1/1/2020
}
\keyword{datasets}
