% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serpico2.R
\docType{package}
\name{serpico2}
\alias{serpico2}
\title{serpico2: Seroprevalence of SARS-CoV-2.}
\description{
This package provides functions to reproduce the analysis from the manuscript "Prevalence of SARS-CoV-2 antibodies in France: results from nationwide serological surveillance" <https://www.medrxiv.org/content/10.1101/2020.10.20.20213116v1>.
}
