% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_stan}
\alias{set_stan}
\title{Set Stan inputs}
\usage{
set_stan(
  df = d0,
  calib = calib_unit_test(),
  modelnb = 6,
  n_iter = 200,
  n_warmup = 150,
  n_chains = 2,
  n_core = 2,
  scale_beta_prior = 2,
  scale_sigma_prior = 1
)
}
\arguments{
\item{df}{sample data}

\item{calib}{named vector of {"ncc", "npp", "TP", "TN"} counts}

\item{modelnb}{for reference in ouput name}

\item{n_iter, n_warmup, n_chains, n_core}{mcmc parameter}
}
\description{
Set Stan inputs
}
