% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate_mock_data}
\alias{simulate_mock_data}
\title{Simulate dataset}
\usage{
simulate_mock_data(
  vars = c("time", "sex", "age", "reg"),
  ntime = 3,
  nsex = 2,
  nage = 9,
  nreg = 17,
  prev = 0.08,
  nsamplecell = 10,
  risk_time = c(0.1, 1, 1.2),
  risk_sex = c(1, 1.2),
  risk_age = c(0.5, rep(1, nage - 1)),
  risk_reg = c(rep(1, nreg - 1), 2),
  calib = calib_unit_test()
)
}
\arguments{
\item{vars}{names of covariates}

\item{ntime, nsex, nage, nreg}{number of covariates categories}

\item{prev}{true prevalence in reference category}

\item{nsamplecell}{number of sample per poststratification * period cell}

\item{risk_time, risk_sex, risk_age, risk_reg}{relative risks ({ntime,nsex,nage,nreg}_vectors)}

\item{calib}{named vector of {"ncc", "npp", "TP", "TN"} counts}
}
\description{
Simulate dataset
}
